requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py@64849ebcc84c38c305f2ade21b6e8212b2e42090
psycopg2-binary==2.9.3
python-dotenv==0.20.0
PyYAML==6.0
SQLAlchemy==1.4.37
requests==2.30.0